Mahmoud 4 commit , work on HomeScreenFragment.kt and add to it NewRecipe button  , also add some drawable design , create new destination called (ResultScreenFragment) to view Results of analysis , work on its design .

+++ app/src/main/java/com/mahmoudbashir/edamamapp/pojo/FOLAC.kt
@@ -2,6 +2,6 @@ package com.mahmoudbashir.edamamapp.pojo
 
 data class FOLAC(
     val label: String,
-    val quantity: Int,
+    val quantity: Double,
     val unit: String
 )
